Move RefSequence to a new refseq module
